Revert "[SWDEV-446215] Update cmake to put test libs in proper lib dir"
This reverts commit 6e01df00ca3b6cb1f4bc156529dfd9ecd0d078de.

Reason for revert: Because the gtest of amdsmi is different to other components so it was installed in a share/amdsmi/lib folder. It cannot be installed in a common folder such as /usr/local/bin or /usr/bin because all other components try to search those folder first.

 

This is breaking ROCmValidationSuite and other tools. Per Wang, Yanyao this should be reverted.

Change-Id: Id61bc6056fe41800e738616f39293e9b8762a377